server: move DnssecSummary directly into parent module
